+++ src/entities/auth/join/types.ts
@@ -1,0 +1,19 @@
+import z from 'zod'
+import { joinSchema } from './joinSchema'
+
+export type Inputs = z.infer<typeof joinSchema>
+
+export interface JoinResponse {
+  msg: string
+  data: string
+}
+
+export interface CreateCodeResponse {
+  msg: string
+  data: string
+}
+
+export interface CheckCodeResponse {
+  mail: string
+  data: string
+}
