+++ src/entities/auth/login/types.ts
@@ -1,0 +1,5 @@
+export interface MyInfo {
+  id: number
+  mail: string
+  nickname: string
+}
