+++ src/entities/project/types.ts
@@ -1,0 +1,9 @@
+export interface Project {
+  id: number
+  title: string
+}
+
+export interface ProjectsRes {
+  msg: string
+  data: Project[]
+}
