+++ src/pages/__root.tsx
@@ -2,6 +2,7 @@ import * as React from 'react'
 import { Outlet, createRootRouteWithContext } from '@tanstack/react-router'
 import { TanStackRouterDevtools } from '@tanstack/react-router-devtools'
 import { type RouterContext } from '~app/router'
+import Sidebar from '~widgets/Sidebar'
 
 // find method to fix this
 export const Route = createRootRouteWithContext<RouterContext>()({
@@ -11,6 +12,7 @@ export const Route = createRootRouteWithContext<RouterContext>()({
 function RootComponent() {
   return (
     <React.Fragment>
+      <Sidebar />
       <Outlet />
       <TanStackRouterDevtools />
     </React.Fragment>
