+++ src/pages/projects/$projectId.tsx
@@ -1,0 +1,9 @@
+import { createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute('/projects/$projectId')({
+  component: RouteComponent,
+})
+
+function RouteComponent() {
+  return <div>Hello "/projects/$projectId"!</div>
+}
